Add once_cell crate and initialize DEFAULT_CONTEXT variable
- Add once_cell crate to Cargo.toml
- Initialize DEFAULT_CONTEXT variable in consts.rs
- Add Context::new() function in context/mod.rs to create a new Context
instance

This commit adds the once_cell crate to the project by updating the
Cargo.toml file. It also initializes the DEFAULT_CONTEXT variable in the
consts.rs file, using the Lazy struct from the once_cell crate. Lastly,
it adds a new function named Context::new() in the context/mod.rs file
to create a new instance of the Context struct.