test: add register_formatter_plugin_succeeds test
This commit adds a new test `register_formatter_plugin_succeeds` to
verify the successful registration of a formatter plugin. The test
registers the provided formatter plugin `TEST_FORMATTER_PLUGIN` and
checks if the plugin is correctly added to the context's list of
formatters. If the registration succeeds, the test passes; otherwise, an error is
returned. The test also defines a mock `TEST_FORMATTER_IN` and
`TEST_FORMATTER_OUT` functions that should never be called.