Refactor Context to expose InterpFnFactory getter
This commit refactors the Context struct in the
lib/src/state/context/mod.rs file to add a new public method
`get_interp_factory()` which returns the `interp_factory` field value.
The addition of this getter allows external code to access the
underlying `interp_factory` value, enhancing the flexibility and
extensibility of the Context struct. This change simplifies the code and
provides a more convenient way to retrieve the InterpFnFactory.